Update stats
Deleting all but two metrics to focus on MVP and getting this through the door.

Slightly improving the sqlite members schema to work with Go's time.Time.

I think that I'm going to trigger the monthly update by hitting an endpoint, so there isn't really a reason to have a CLI anymore.